Removed the triplet carbene groups from forbiddenStructures.py
I had transition states for triplet radicals, and these were not
used to train the groups (also had triplet groups).

There is an issue with singlet and triplet carbenes in RMG,
and I think it is about not having separate kinetic data for
singlet and triplet states. If that is the case, this branch
would only help.