Add GAV for oxitene and oxerene ring groups
GAV's derived from just the molecules themselves which I found in the
DFT_QCI library. Adding the GAV values makes it so that substituted
molecules will be much more accurate during model generation.

Admittedly the method for deriving the GAV is not great because of the
minimal training set, but it will probably still be much better than it
was.